requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
billiard==3.5.0.5
celery==4.2.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-summernote==0.8.11.6
idna==2.9
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.1.0
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
wrapt==1.11.2
